18/12

+++ database/migrations/2023_12_08_073132_create_products_table.php
@@ -18,7 +18,7 @@ return new class extends Migration {
             $table->text("description");
             $table->text("nutrition_detail");
             $table->mediumInteger("stock_quantity")->default(0);
-            $table->mediumInteger("store_quantity")->default(0);
+            $table->float("store_quantity", 8, 2)->default(0);
             $table->unsignedFloat("price", 8, 2);
             $table->tinyInteger("star")->default(5);
             $table->tinyInteger("status")->default(1)->comment("1:Show - 2:Hidden");
